requirements.txt
anthropic>=0.25.0
bcrypt>=4.1.2
cryptography>=42.0.8
extra-streamlit-components>=0.1.71
openai>=1.12.0
PyJWT>=2.8.0
python-dotenv>=1.0.1
streamlit>=1.44.0
